Fix two cancelling errors on math.cpp and test/ballistics.test.cpp
In every test of the Ballistics class,
the rotation from the pair p1 and p2 to the pair q1 and q2
was reversed. To go from p1 to q1 there was necessary
a negative rotation instead of the positive one that was present
on the test. This could mess with the axis computation.

However, the compute_rotation_axis function in math.cpp
was actually using a left-handed coordinate system;
so, that function rotated the vector to the wrong side,
which was actually the right side
because the vectors were in the wrong order.

This patch fixes both errors.
compute_rotation_axis now uses the right-hand rule,
and the angles in the test cases correspont do the actual rotations
of the measurements p1, p2, q1, q2.