ballistics.cpp: Reorder computation of angles in Ballistics::align
The previous computation of rotation_data::secondary used projections
to compute it separately from rotation_data::main. This is wrong;
the secondary rotation plane itself rotates around the main axis,
so projections change the angle.

Thus, now we first compute rotation_data::main, then rotate _front
and _left to align the secondary rotation plane with the vector p (the
last measurement) and finally compute the angle between the new _front
and p. This avoids the angle distortion of the projection.